Reformatted redux
Separated redux by entity for scalability. Moved Provider wrapper to index.js. Used catch instead of onFail callback for fetch.
