+++ src/App.js
@@ -1,14 +1,8 @@
 import React from 'react'
-import { Provider } from 'react-redux'
-import store from './store';
 import AppContainer from './AppContainer'
 
 function App() {
-  return (
-    <Provider store={store}>
-      <AppContainer/>
-    </Provider>
-  )
+  return <AppContainer/>
 }
 
 export default App
