--- src/reducers/index.js
@@ -1,10 +1,0 @@
-import { combineReducers } from 'redux';
-import authReducer from './authReducer';
-import businessReducer from './businessReducer';
-
-const rootReducer = combineReducers({
-  auth: authReducer,
-  businesses: businessReducer
-});
-
-export default rootReducer;
