+++ src/store/businesses/actionTypes.js
@@ -1,0 +1,3 @@
+export const SET_ERROR = 'SET_ERROR';
+
+export const GET_BUSINESSES = 'GET_BUSINESSES';
