+++ src/store/businesses/actions.js
@@ -1,6 +1,4 @@
-export const SET_ERROR = 'SET_ERROR';
-
-export const GET_BUSINESSES = 'GET_BUSINESSES';
+import { GET_BUSINESSES, SET_ERROR } from './actionTypes'
 
 export const setError = (message) => {
   return {
@@ -12,23 +10,19 @@ export const setError = (message) => {
 export const getAllBusinesses = () => {
   return (dispatch) => {
     fetch('https://warm-cliffs-11938.herokuapp.com/api/businesses')
-      .then(
-        res => res.json(),
-        err => {
-          dispatch(setError(err.message)); // Error w/ api call
-        })
+      .then(res => res.json())
       .then(businesses => {
           dispatch({
             type: GET_BUSINESSES,
             payload: businesses
           })
-        },
-        err => {
-          dispatch({
-            type: SET_ERROR,
-            payload: err.message // Error parsing response as json
-          })
         }
       )
+      .catch(err => {
+        dispatch({
+          type: SET_ERROR,
+          payload: err.message // Error parsing response as json
+        })
+      })
   }
 }
