+++ src/store/index.js
@@ -1,6 +1,12 @@
-import { createStore, applyMiddleware } from 'redux';
+import { createStore, combineReducers, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
-import rootReducer from '../reducers';
+import authReducer from '../store/users/reducer';
+import businessReducer from '../store/businesses/reducer';
+
+const rootReducer = combineReducers({
+  auth: authReducer,
+  businesses: businessReducer
+});
 
 const store = createStore(rootReducer, applyMiddleware(thunk));
