+++ src/store/users/actionTypes.js
@@ -1,0 +1,1 @@
+export const SET_AUTH_STATUS = 'SET_AUTH_STATUS';
