+++ src/store/users/actions.js
@@ -1,0 +1,8 @@
+import { SET_AUTH_STATUS } from './actionTypes'
+
+export const setAuthStatus = (status) => {
+  return {
+    type: SET_AUTH_STATUS,
+    payload: status
+  }
+};
